code changes: toggle password, forgot password and sign up links, dynamic error messages, updated src file structure
