+++ llm_bioinformatics_research/src/LoginScreen/LoginScreen.js
@@ -1,0 +1,140 @@
+import React, { useState } from 'react';
+import { TextField, Button, Container, Typography, Box, IconButton, Link } from '@mui/material';
+import { Visibility, VisibilityOff } from '@mui/icons-material';
+
+function LoginScreen() {
+  const [email, setEmail] = useState('');
+  const [password, setPassword] = useState('');
+  const [emailError, setEmailError] = useState('');
+  const [passwordError, setPasswordError] = useState('');
+  const [passwordVisible, setPasswordVisible] = useState(false);
+
+  const validateEmail = (email) => {
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    if (!emailRegex.test(email)) {
+      setEmailError('Please enter a valid email address.');
+      return false;
+    }
+    setEmailError('');
+    return true;
+  };
+
+  const validatePassword = (password) => {
+    if (password.length < 8) {
+      setPasswordError('Password must be at least 8 characters long.');
+      return false;
+    }
+    if (!/[A-Z]/.test(password)) {
+      setPasswordError('Password must include at least one uppercase letter.');
+      return false;
+    }
+    if (!/[a-z]/.test(password)) {
+      setPasswordError('Password must include at least one lowercase letter.');
+      return false;
+    }
+    if (!/\d/.test(password)) {
+      setPasswordError('Password must include at least one number.');
+      return false;
+    }
+    if (!/[@$!%*#?&]/.test(password)) {
+      setPasswordError('Password must include at least one special character.');
+      return false;
+    }
+    setPasswordError('');
+    return true;
+  };
+
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    const isEmailValid = validateEmail(email);
+    const isPasswordValid = validatePassword(password);
+
+    if (isEmailValid && isPasswordValid) {
+      console.log('Form submitted');
+    }
+  };
+
+  return (
+    <Container maxWidth="sm">
+      <Box
+        sx={{
+          display: 'flex',
+          flexDirection: 'column',
+          alignItems: 'center',
+          justifyContent: 'center',
+          height: '100vh',
+        }}
+      >
+        <Typography variant="h4" gutterBottom data-testid="login-heading">
+          Login
+        </Typography>
+        <form onSubmit={handleSubmit} noValidate>
+        <TextField
+            id="email"
+            label="Email"
+            type="email"
+            variant="outlined"
+            fullWidth
+            margin="normal"
+            value={email}
+            onChange={(e) => setEmail(e.target.value)}
+            error={!!emailError}
+            helperText={emailError}
+            inputProps={{
+              'data-testid': 'email-input',
+            }}
+          />
+          <TextField
+            id="password"
+            label="Password"
+            type={passwordVisible ? 'text' : 'password'}
+            variant="outlined"
+            fullWidth
+            margin="normal"
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+            error={passwordError !== ''}
+            helperText={passwordError}
+            InputProps={{
+              endAdornment: (
+                <IconButton
+                  aria-label="toggle password visibility"
+                  onClick={() => setPasswordVisible(!passwordVisible)}
+                  edge="end"
+                >
+                  {passwordVisible ? <VisibilityOff /> : <Visibility />}
+                </IconButton>
+              ),
+            }}
+            // Add testid to the TextField component
+            inputProps={{
+              'data-testid': 'password-input',
+            }}
+          />
+          <Button
+            type="submit"
+            variant="contained"
+            color="primary"
+            fullWidth
+            sx={{ mt: 2 }}
+            data-testid="login-button"
+          >
+            Login
+          </Button>
+          <Box sx={{ mt: 2, textAlign: 'center' }}>
+            <Link href="#" variant="body2">
+              Forgot Password?
+            </Link>
+          </Box>
+          <Box sx={{ mt: 1, textAlign: 'center' }}>
+            <Link href="#" variant="body2">
+              Sign Up
+            </Link>
+          </Box>
+        </form>
+      </Box>
+    </Container>
+  );
+}
+
+export default LoginScreen;
