+++ llm_bioinformatics_research/src/Testing/App.test.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import '@testing-library/jest-dom';
+import App from '../App';
+
+test('renders login screen', () => {
+  render(<App />);
+  const loginButton = screen.getByTestId('login-button'); 
+  expect(loginButton).toBeInTheDocument();
+});
+
