+++ llm_bioinformatics_research/src/Testing/LoginScreen.test.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import LoginScreen from './LoginScreen';
+import LoginScreen from '../LoginScreen/LoginScreen';
 
 console.log = jest.fn();
 
 describe('LoginScreen Component', () => {
   test('should show error for invalid email format', () => {
     render(<LoginScreen />);
-    const emailInput = screen.getByLabelText(/email/i);
+    const emailInput = screen.getByTestId('email-input');
     const submitButton = screen.getByRole('button', { name: /login/i });
 
     fireEvent.change(emailInput, { target: { value: 'invalidemail' } });
@@ -19,7 +19,7 @@ describe('LoginScreen Component', () => {
 
   test('should show no error for valid email format', () => {
     render(<LoginScreen />);
-    const emailInput = screen.getByLabelText(/email/i);
+    const emailInput = screen.getByTestId('email-input');
     const submitButton = screen.getByRole('button', { name: /login/i });
 
     fireEvent.change(emailInput, { target: { value: 'validemail@example.com' } });
@@ -30,7 +30,7 @@ describe('LoginScreen Component', () => {
 
   test('should show error when email is empty', () => {
     render(<LoginScreen />);
-    const emailInput = screen.getByLabelText(/email/i);
+    const emailInput = screen.getByTestId('email-input');
     const submitButton = screen.getByRole('button', { name: /login/i });
 
     fireEvent.change(emailInput, { target: { value: '' } });
@@ -41,41 +41,74 @@ describe('LoginScreen Component', () => {
 
   test('should show error for password that is too short', () => {
     render(<LoginScreen />);
-    const passwordInput = screen.getByLabelText(/password/i);
+    const passwordInput = screen.getByTestId('password-input');
     const submitButton = screen.getByRole('button', { name: /login/i });
 
     fireEvent.change(passwordInput, { target: { value: 'Ab1!' } });
     fireEvent.click(submitButton);
 
-    expect(screen.getByText(/password must be at least 8 characters long/i)).toBeInTheDocument();
+    expect(screen.getByText(/at least 8 characters long/i)).toBeInTheDocument();
+  });
+
+  test('should show error for missing uppercase letter in password', () => {
+    render(<LoginScreen />);
+    const passwordInput = screen.getByTestId('password-input');
+    const submitButton = screen.getByRole('button', { name: /login/i });
+
+    fireEvent.change(passwordInput, { target: { value: 'abcdef12!' } });
+    fireEvent.click(submitButton);
+
+    expect(screen.getByText(/at least one uppercase letter./i)).toBeInTheDocument();
+  });
+
+  test('should show error for missing lowercase letter in password', () => {
+    render(<LoginScreen />);
+    const passwordInput = screen.getByTestId('password-input');
+    const submitButton = screen.getByRole('button', { name: /login/i });
+
+    fireEvent.change(passwordInput, { target: { value: 'ABCDEF12!' } });
+    fireEvent.click(submitButton);
+
+    expect(screen.getByText(/at least one lowercase letter./i)).toBeInTheDocument();
+  });
+
+  test('should show error for missing number in password', () => {
+    render(<LoginScreen />);
+    const passwordInput = screen.getByTestId('password-input');
+    const submitButton = screen.getByRole('button', { name: /login/i });
+
+    fireEvent.change(passwordInput, { target: { value: 'Abcdefgh!' } });
+    fireEvent.click(submitButton);
+
+    expect(screen.getByText(/at least one number./i)).toBeInTheDocument();
   });
 
   test('should show error for password missing special character', () => {
     render(<LoginScreen />);
-    const passwordInput = screen.getByLabelText(/password/i);
+    const passwordInput = screen.getByTestId('password-input');
     const submitButton = screen.getByRole('button', { name: /login/i });
-  
+
     fireEvent.change(passwordInput, { target: { value: 'Abcdef12' } });
     fireEvent.click(submitButton);
-  
-    expect(screen.getByText(/password must be at least 8 characters long/i)).toBeInTheDocument();
+
+    expect(screen.getByText(/at least one special character/i)).toBeInTheDocument();
   });
 
   test('should show no error for valid password', () => {
     render(<LoginScreen />);
-    const passwordInput = screen.getByLabelText(/password/i);
+    const passwordInput = screen.getByTestId('password-input');
     const submitButton = screen.getByRole('button', { name: /login/i });
-  
+
     fireEvent.change(passwordInput, { target: { value: 'Abcdef12!' } });
     fireEvent.click(submitButton);
-  
+
     expect(screen.queryByText(/password must be at least 8 characters long./i)).not.toBeInTheDocument();
   });
 
   test('should call console.log on valid submission', () => {
     render(<LoginScreen />);
-    const emailInput = screen.getByLabelText(/email/i);
-    const passwordInput = screen.getByLabelText(/password/i);
+    const emailInput = screen.getByTestId('email-input');
+    const passwordInput = screen.getByTestId('password-input');
     const submitButton = screen.getByRole('button', { name: /login/i });
 
     fireEvent.change(emailInput, { target: { value: 'validemail@example.com' } });
@@ -87,8 +120,8 @@ describe('LoginScreen Component', () => {
 
   test('should not call console.log on invalid submission', () => {
     render(<LoginScreen />);
-    const emailInput = screen.getByLabelText(/email/i);
-    const passwordInput = screen.getByLabelText(/password/i);
+    const emailInput = screen.getByTestId('email-input');
+    const passwordInput = screen.getByTestId('password-input');
     const submitButton = screen.getByRole('button', { name: /login/i });
 
     fireEvent.change(emailInput, { target: { value: 'invalidemail' } });
@@ -98,4 +131,3 @@ describe('LoginScreen Component', () => {
     expect(console.log).not.toHaveBeenCalled();
   });
 });
-
